Problem 70 solution
Also made a minor update to Problem 69 and a number of updates to DivisorsUtil. While working on my initial solutions to Problem 70 I found some of the logic was incorrect in DivisorsUtil. I updated it, as well as added 2 getPrimeDivisors(..) methods. Added proper unit tests for DivisorsUtil so I know the logic works as expected, which I should have done initially.